Configuration endpoint for local configuration
This makes it possible to read and write configuration (like ledMode and
mqtt url) and trigger different calibrations locally.

It is a draft meant to spark a discussion, because while it does update
the configuration on the device, the setting only last until they are
updated (and overwritten) from the cloud (for an average of 7.5 seconds).