Add combined `hybrid` led bar mode
This commit adds an additional led mode for the AirGradient One devices.
Its called `hybrid` and it combines the CO2 and PM2.5 led modes
displaying the worst value of both.

This allows the user to immediately evaluate the current air quality
with a single glance. CO2 and PM2.5 can be checked combined directly from it.